Moved MainViewModel to features/navigation/presentation/viewmodel
