+++ app/src/main/java/com/nailorsh/repeton/MainActivity.kt
@@ -1,16 +1,9 @@
 package com.nailorsh.repeton
 
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import com.nailorsh.repeton.MainActivity.Companion.mainActivity
-import com.nailorsh.repeton.core.ui.theme.RepetonTheme
-import com.nailorsh.repeton.features.navigation.ui.RepetonApp
+import com.nailorsh.repeton.features.navigation.presentation.ui.RepetonApp
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -24,20 +17,7 @@ class MainActivity : ComponentActivity() {
         mainActivity = this
         super.onCreate(savedInstanceState)
         setContent {
-            val viewModel: MainViewModel = hiltViewModel()
-            val mainUiState = viewModel.uiState.collectAsState()
-            when (val currentState = mainUiState.value) {
-                is MainUiState.Error -> {}
-                is MainUiState.Loading -> {}
-
-                is MainUiState.Success -> {
-                    Log.d("MAIN", currentState.isDarkThemeEnabled.toString())
-                    RepetonTheme(darkTheme = currentState.isDarkThemeEnabled) {
-                        RepetonApp()
-                    }
-                }
-            }
-
+            RepetonApp()
         }
     }
