+++ app/src/main/java/com/nailorsh/repeton/features/navigation/presentation/ui/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.nailorsh.repeton.features.navigation.ui
+package com.nailorsh.repeton.features.navigation.presentation.ui
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
